requirements.txt
transformers==4.36.2
spacy==3.7.2
python-docx==1.0.1
nltk==3.8.1
tqdm==4.66.1 